Honor scheme when rebuilding URL to retrieve access token from Guisso
If the `host` parameter given to Api.from_oauth_client() is not a full URL (ie.
does not include the protocol scheme) but the given port is 443 because `https`
is true, the code was reconstructing the original URL as http://host:443 which
yields a (generally) incorrect value for the request app scope. So, if no scheme
is given, infer it from the value of the `https` parameter.